requirements.txt
fastapi==0.75.0
keras==2.8.0
Keras-Preprocessing==1.1.2
Markdown==3.3.6
matplotlib==3.5.1
numpy==1.21.5
pandas==1.3.5
Pillow==9.0.1
python-multipart==0.0.5
requests
tensorflow==2.8.0
tensorflow-datasets==4.5.2
tensorflow-serving-api==2.5.0
uvicorn==0.17.6
unsync